Move frustration observer
